requirements.txt
termcolor
llama-cpp-python
